Relocate AnimationTabRefreshLogic out of MainStateAnimationPlugin, part of #3866
Moves the animation tab's pure reload/reselect decision logic (already
WPF-free and already pinned) from the WPF-hosted MainStateAnimationPlugin
into a new public static class in headless Gum.Presentation, so it's
reachable without the WPF plugin project. Production call sites and the
existing pinning tests are retargeted; no behavior change.

Boyscout: dropped the now-dead InternalsVisibleTo("GumToolUnitTests") on
StateAnimationPlugin.csproj's AssemblyInfo, since these relocated members
were its only reason for existing.
