+++ Gum/StateAnimationPlugin/MainStateAnimationPlugin.cs
@@ -291,11 +291,11 @@ public class MainStateAnimationPlugin : PluginBase, IAnimationUndoProvider
             ? _animationFilePathService.GetAbsoluteAnimationFileNameFor(selectedElement)
             : null;
 
-        if (ShouldReloadAnimationsForChangedFile(filePath, selectedElementAnimationFile))
+        if (AnimationTabRefreshLogic.ShouldReloadAnimationsForChangedFile(filePath, selectedElementAnimationFile))
         {
             // Capture the selection before the rebuild replaces every animation/keyframe instance, then
             // reapply it so the external edit doesn't deselect the user's animation + keyframe (#3410).
-            var selection = CaptureAnimationSelection(_viewModel);
+            var selection = AnimationTabRefreshLogic.CaptureAnimationSelection(_viewModel);
 
             // forceReload bypasses CreateViewModel's same-element early-out: the external edit doesn't
             // change the selection, so without forcing it the stale view model would survive and the
@@ -305,26 +305,9 @@ public class MainStateAnimationPlugin : PluginBase, IAnimationUndoProvider
 
             if (_viewModel != null)
             {
-                RestoreAnimationSelection(_viewModel, selection);
+                AnimationTabRefreshLogic.RestoreAnimationSelection(_viewModel, selection);
             }
         }
-    }
-
-    /// <summary>
-    /// Decides whether an on-disk file change should live-reload the Animations tab (issue #3410):
-    /// true only when <paramref name="changedFile"/> is a <c>.ganx</c> and is the selected element's
-    /// own animation sidecar (<paramref name="selectedElementAnimationFile"/>). Other elements' .ganx
-    /// files reload lazily when that element is next selected, so they are ignored here.
-    /// </summary>
-    internal static bool ShouldReloadAnimationsForChangedFile(FilePath changedFile,
-        FilePath? selectedElementAnimationFile)
-    {
-        if (changedFile.Extension != "ganx")
-        {
-            return false;
-        }
-
-        return selectedElementAnimationFile != null && changedFile == selectedElementAnimationFile;
     }
 
     private void HandleElementSelected(ElementSave? element)
@@ -427,24 +410,12 @@ public class MainStateAnimationPlugin : PluginBase, IAnimationUndoProvider
         var element = _selectedState.SelectedElement;
         if (element != null && _viewModel != null)
         {
-            RefreshAfterStateRename(_renameManager, _viewModel, element, stateSave, oldName);
+            AnimationTabRefreshLogic.RefreshAfterStateRename(_renameManager, _viewModel, element, stateSave, oldName);
         }
 
         // Refresh available states / DataContext and re-broadcast. RefreshErrors runs again here,
         // but it is idempotent and rename is not a hot path.
         RefreshViewModel();
-    }
-
-    /// <summary>
-    /// Applies a state rename to <paramref name="viewModel"/>'s keyframe references and then
-    /// recomputes errors, in that order. The order matters: recomputing before the rewrite leaves a
-    /// stale "references a missing state" error on the renamed keyframe (issue #3383).
-    /// </summary>
-    internal static void RefreshAfterStateRename(IRenameManager renameManager,
-        ElementAnimationsViewModel viewModel, ElementSave element, StateSave stateSave, string oldName)
-    {
-        renameManager.HandleRename(stateSave, oldName, viewModel);
-        viewModel.RefreshErrors(element);
     }
 
     private void HandleAfterUndo()
@@ -454,8 +425,8 @@ public class MainStateAnimationPlugin : PluginBase, IAnimationUndoProvider
         // user's animation + keyframe selection survives undo/redo — mirroring element-undo's state
         // selection restore (#3406). Index + count are captured alongside the keyframe so the reselect
         // can fall back to position when the undo reverted the selected keyframe's own value (see
-        // RestoreAnimationSelection).
-        var selection = CaptureAnimationSelection(_viewModel);
+        // AnimationTabRefreshLogic.RestoreAnimationSelection).
+        var selection = AnimationTabRefreshLogic.CaptureAnimationSelection(_viewModel);
 
         // Repaint the tab from the just-restored .ganx, then re-arm undo recording. The flag spanned
         // the .ganx write (ApplyAnimations) through this repaint so neither flushed a spurious undo.
@@ -466,122 +437,10 @@ public class MainStateAnimationPlugin : PluginBase, IAnimationUndoProvider
 
         if (_viewModel != null)
         {
-            RestoreAnimationSelection(_viewModel, selection);
+            AnimationTabRefreshLogic.RestoreAnimationSelection(_viewModel, selection);
         }
 
         _isApplyingUndo = false;
-    }
-
-    /// <summary>
-    /// The animation/keyframe selection captured before a forced view-model rebuild (undo/redo or an
-    /// external .ganx reload). Carries both the identity (animation name + keyframe content) and the
-    /// position (animation/keyframe index + sibling counts) so <see cref="RestoreAnimationSelection"/>
-    /// can fall back to the slot when the identity changed — e.g. the selected animation was renamed.
-    /// </summary>
-    internal readonly record struct AnimationSelectionState(
-        string? AnimationName,
-        AnimatedKeyframeViewModel? Keyframe,
-        int KeyframeIndex,
-        int KeyframeCount,
-        int AnimationIndex = -1,
-        int AnimationCount = 0);
-
-    /// <summary>
-    /// Snapshots the currently-selected animation and keyframe (identity + position) so the selection
-    /// can be reapplied after a forced view-model rebuild replaces every animation/keyframe instance.
-    /// Pairs with <see cref="RestoreAnimationSelection"/>.
-    /// </summary>
-    internal static AnimationSelectionState CaptureAnimationSelection(ElementAnimationsViewModel? viewModel)
-    {
-        var selectedAnimation = viewModel?.SelectedAnimation;
-        var selectedKeyframe = selectedAnimation?.SelectedKeyframe;
-        return new AnimationSelectionState(
-            selectedAnimation?.Name,
-            selectedKeyframe,
-            selectedKeyframe != null ? selectedAnimation!.Keyframes.IndexOf(selectedKeyframe) : -1,
-            selectedAnimation?.Keyframes.Count ?? 0,
-            selectedAnimation != null && viewModel != null ? viewModel.Animations.IndexOf(selectedAnimation) : -1,
-            viewModel?.Animations.Count ?? 0);
-    }
-
-    /// <summary>
-    /// Reselects, on a freshly-rebuilt <paramref name="viewModel"/>, the animation and keyframe captured
-    /// in <paramref name="selection"/>. The animation is matched by name; if that fails because it was
-    /// renamed (e.g. an external .ganx edit), it falls back to the captured animation index when the
-    /// animation count is unchanged. The keyframe is then matched by content; if that fails because the
-    /// undo reverted the selected keyframe's <em>own</em> value (e.g. its time), it falls back to the
-    /// captured index — but only when the keyframe count is unchanged, so an add/delete (which
-    /// changes the count) drops that selection rather than grabbing a neighbor. Returns the matched
-    /// keyframe (or null). Best-effort, mirroring element-undo's silent selection drop when the selected
-    /// object no longer exists.
-    /// </summary>
-    /// <remarks>
-    /// The keyframe is selected on the animation <em>before</em> the animation is made the active
-    /// SelectedAnimation. That ordering matters: setting SelectedAnimation rebinds the keyframes
-    /// ListBox's ItemsSource, and the ListBox initializes its SelectedItem from the (two-way) bound
-    /// SelectedKeyframe at bind time. If SelectedKeyframe is still null then, the ListBox settles on no
-    /// selection and a later assignment gets reset; pre-setting it means the ListBox binds straight to
-    /// the right, already-present keyframe — so the selection (and the right-side property panel) sticks
-    /// without any dispatcher timing games (#3406).
-    /// </remarks>
-    internal static AnimatedKeyframeViewModel? RestoreAnimationSelection(ElementAnimationsViewModel viewModel,
-        AnimationSelectionState selection)
-    {
-        if (selection.AnimationName == null)
-        {
-            return null;
-        }
-
-        var animation = viewModel.Animations.FirstOrDefault(item => item.Name == selection.AnimationName);
-
-        // The selected animation may have been renamed by an external .ganx edit, so the captured name
-        // matches nothing. Fall back to the captured slot when the animation count is unchanged, keeping
-        // the same row selected through the rename (#3410). A count change means an add/delete, where
-        // grabbing a neighbor by index would be wrong, so the selection drops instead.
-        if (animation == null
-            && viewModel.Animations.Count == selection.AnimationCount
-            && selection.AnimationIndex >= 0
-            && selection.AnimationIndex < viewModel.Animations.Count)
-        {
-            animation = viewModel.Animations[selection.AnimationIndex];
-        }
-
-        if (animation == null)
-        {
-            return null;
-        }
-
-        AnimatedKeyframeViewModel? matched = null;
-        if (selection.Keyframe != null)
-        {
-            matched = animation.Keyframes.FirstOrDefault(item => AreSameKeyframe(item, selection.Keyframe));
-
-            if (matched == null
-                && animation.Keyframes.Count == selection.KeyframeCount
-                && selection.KeyframeIndex >= 0
-                && selection.KeyframeIndex < animation.Keyframes.Count)
-            {
-                matched = animation.Keyframes[selection.KeyframeIndex];
-            }
-        }
-
-        animation.SelectedKeyframe = matched;
-        viewModel.SelectedAnimation = animation;
-
-        return matched;
-    }
-
-    /// <summary>
-    /// Identity match for a keyframe across a view-model rebuild: same discriminator (state /
-    /// sub-animation / event name) and time. Used to reselect the previously-selected keyframe on the
-    /// new instances after an undo/redo reload.
-    /// </summary>
-    private static bool AreSameKeyframe(AnimatedKeyframeViewModel first, AnimatedKeyframeViewModel second)
-    {
-        return first.StateName == second.StateName
-            && first.AnimationName == second.AnimationName
-            && first.EventName == second.EventName
-            && first.Time == second.Time;
     }
 
     private void HandleStateAdd(StateSave state)
@@ -775,64 +634,7 @@ public class MainStateAnimationPlugin : PluginBase, IAnimationUndoProvider
 
     private void RefreshAvailableStates()
     {
-        AvailableStates.ReplaceWith(GetAvailableStates(_selectedState.SelectedElement, _viewModel));
-    }
-
-    /// <summary>
-    /// Builds the state names shown in each keyframe's state ComboBox: every state on the element,
-    /// plus any state still referenced by a keyframe even though it no longer exists on the element.
-    /// Keeping a referenced-but-missing state in the list matters because the ComboBox is editable
-    /// with its Text bound to the keyframe's StateName — if the referenced item left the ItemsSource,
-    /// the ComboBox would coerce its Text (and thus StateName) to empty, collapsing a broken state
-    /// keyframe into an event (issue #3392). The keyframe stays flagged as broken via
-    /// HasValidState/RefreshErrors, which is computed against the element, not this list.
-    /// </summary>
-    internal static List<string> GetAvailableStates(ElementSave? element, ElementAnimationsViewModel? viewModel)
-    {
-        var states = new List<string>();
-
-        if (element != null)
-        {
-            states.AddRange(element.States.Select(item => item.Name));
-
-            foreach (var category in element.Categories)
-            {
-                states.AddRange(category.States.Select(item => category.Name + "/" + item.Name));
-            }
-        }
-
-        // Keep any state still referenced by a keyframe even though it no longer exists on the
-        // element (e.g. its category was just deleted). The editable state ComboBox binds its Text to
-        // the keyframe's StateName; if the referenced item left this list it would coerce StateName to
-        // empty, collapsing the broken state keyframe into an event (issue #3392).
-        if (viewModel != null)
-        {
-            foreach (var animation in viewModel.Animations)
-            {
-                foreach (var keyframe in animation.Keyframes)
-                {
-                    if (!string.IsNullOrEmpty(keyframe.StateName) && !states.Contains(keyframe.StateName))
-                    {
-                        states.Add(keyframe.StateName);
-                    }
-                }
-            }
-        }
-
-        return states;
-    }
-
-    /// <summary>
-    /// Decides whether <see cref="CreateViewModel"/> should rebuild the view model from the element's
-    /// .ganx. Normally the view model is only reloaded when the selected element changes; an in-place
-    /// undo/redo restores the .ganx without changing the selection, so the after-undo path passes
-    /// <paramref name="forceReload"/> to repaint the tab immediately rather than keeping the stale view
-    /// model until the element is reselected (#3406).
-    /// </summary>
-    internal static bool ShouldReloadViewModel(ElementSave? currentlyReferencedElement,
-        ElementSave? selectedElement, bool forceReload)
-    {
-        return currentlyReferencedElement != selectedElement || forceReload;
+        AvailableStates.ReplaceWith(AnimationTabRefreshLogic.GetAvailableStates(_selectedState.SelectedElement, _viewModel));
     }
 
     private void CreateViewModel(bool forceReload = false)
@@ -845,7 +647,7 @@ public class MainStateAnimationPlugin : PluginBase, IAnimationUndoProvider
 
         var element = _selectedState.SelectedElement;
 
-        if (ShouldReloadViewModel(currentlyReferencedElement, element, forceReload))
+        if (AnimationTabRefreshLogic.ShouldReloadViewModel(currentlyReferencedElement, element, forceReload))
         {
             if (_projectState.GumProjectSave?.FullFileName == null)
             {
